feat: Added sort options and sort view

+++ app/src/main/java/com/pkg/noteapp/model/local/NoteEntity.kt
@@ -10,5 +10,5 @@ data class NoteEntity(
     val title: String,
     val description: String?,
     val dateTime: Long,
-    val color: Int,
+    val color: Long,
 )
